✨ Konamiコマンドをイースターエッグ
